requirements.txt
cffi==1.14.6
cryptography==3.4.8
cycler==0.10.0
dnspython==2.1.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
